split graphql
