+++ src/server/graphql/mutations/toggleCharacter.js
