+++ src/server/middlewares/graphql.js
@@ -1,0 +1,7 @@
+import graphql from 'express-graphql';
+import schema from '../graphql/schema.js';
+
+export default graphql({
+  schema: schema,
+  graphiql: true
+});
